requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.11.0
joblib>=1.3.0
python-dateutil>=2.9.0
